dsp/window: refactor Gaussian and Tukey windows to be each a single type
